backend/api/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
distro==1.9.0
Django==5.0.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.4
mysql-connector-python==8.3.0
numpy==1.26.4
openai==1.13.3
pandas==2.2.1
protobuf==3.20.3
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.0.3
